+++ CHANGES.md
@@ -39,6 +39,7 @@
     - Gestionnaire de fichiers, modifications des paramètres des miniatures.
     - Suppression d'options inutiles dans l'édition d'une page de type de barre latérale.
     - Filtrage du nom des pages dans la fenêtre d'édition des pages.
+    - Format de date dans le module Blog
     
 ## version 9.2.12
 - Modifications